pyproject.toml
[tool.poetry]
name = "news-prediction"
version = "0.1.0"
description = ""
authors = ["Daniil <ksusonic@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
black = "^23.12.1"
ruff = "^0.1.9"
pytest = "^7.4.3"
pre-commit = "^3.6.0"
matplotlib = "^3.8.2"
numpy = ">=1.22,<1.24"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
pandas = "^2.1.4"
mlflow = "^2.9.2"
tensorflow = "2.12.0"
wordcloud = "^1.9.3"
scikit-learn = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120

[tool.black]
skip-string-normalization = true
line-length = 120

[tool.ruff]
# Same as Black.
line-length = 120

[tool.ruff.isort]
order-by-type = true
relative-imports-order = "closest-to-furthest"
extra-standard-library = ["typing"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
known-first-party = []
